+++ spec/rubocop/cop/style/file_name_spec.rb
@@ -79,6 +79,14 @@ describe RuboCop::Cop::Style::FileName do
     end
   end
 
+  context 'with snake_case names which use ? and !' do
+    let(:filename) { 'some/dir/file?!.rb' }
+
+    it 'does not report an offense' do
+      expect(cop.offenses).to be_empty
+    end
+  end
+
   context 'with non-snake-case file names with a shebang' do
     let(:filename) { '/some/dir/test-case' }
     let(:source) { ['#!/usr/bin/env ruby', 'print 1'] }